+++ app/Http/Controllers/ResourceTypeController.php
@@ -7,7 +7,7 @@ use App\Http\Parameters\Route\Validate;
 use App\Models\ResourceType;
 use App\Transformers\ResourceType as ResourceTypeTransformer;
 use App\Utilities\Request as UtilityRequest;
-use App\Validators\ResourceType as ResourceTypeValidator;
+use App\Http\Parameters\Request\Validators\ResourceType as ResourceTypeValidator;
 use Exception;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;